PostgreSQL extracted database/requirements.txt
requests
psycopg2-binary
requests
python-dotenv
geoip2
google-cloud-sql-python-connector
SQLAlchemy
PyMySQL
